LayoutTests/fast/scrolling/ios/programmatic-scroll-while-zoomed.html

<!DOCTYPE html><!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body {
            overflow: hidden;
            height: 1000px;
        }

        .contents {
            width: 100%;
            height: 100%;
        }
        
        /* Prevent output from affecting the document size */
        #console {
            height: 10em;
            overflow: clip;
        }
    </style>
    <script src="../../../resources/ui-helper.js"></script>
    <script src="../../../resources/js-test.js"></script>
    <script>
        jsTestIsAsync = true;

        async function runTest() {
            if (!window.testRunner)
                return;

            await UIHelper.zoomToScale(1.5);
            await UIHelper.renderingUpdate();
            
            shouldBe('window.pageXOffset', '65');
            shouldBe('window.pageYOffset', '133');

            shouldBe('window.visualViewport.pageLeft', '65');
            shouldBe('window.visualViewport.pageTop', '133');

            await Promise.all([UIHelper.ensureVisibleContentRectUpdate(), UIHelper.ensureStablePresentationUpdate()]);

            debug('');
            debug('scrolling to 0,0');
            window.scrollTo(0, 0);

            await Promise.all([UIHelper.ensureVisibleContentRectUpdate(), UIHelper.ensureStablePresentationUpdate()]);

            shouldBe('window.pageXOffset', '65');
            shouldBe('window.pageYOffset', '133');

            shouldBe('window.visualViewport.pageLeft', '65');
            shouldBe('window.visualViewport.pageTop', '133');

            finishJSTest();
        }

        window.addEventListener('load', runTest, false);
    </script>
</head>
<body>
    <div class="contents"></div>
    <div id="console"></div>
</body>
</html>
