LayoutTests/fast/scrolling/ios/programmatic-scroll-while-zoomed-expected.txt

PASS window.pageXOffset is 65
PASS window.pageYOffset is 133
PASS window.visualViewport.pageLeft is 65
PASS window.visualViewport.pageTop is 133

scrolling to 0,0
PASS window.pageXOffset is 65
PASS window.pageYOffset is 133
PASS window.visualViewport.pageLeft is 65
PASS window.visualViewport.pageTop is 133
PASS successfullyParsed is true

TEST COMPLETE
